Adds comments table and model
Comments can belong to either a post (if they are a top-level comment)
or another comment (in the case that they are a nested comment.)  By
making the parent polymorphic, and adding the appropriate has_many
association to comment and post, comments can now belong to either
entity.

Perhaps a better solution would be to always associate a comment with a
post, and then optionally a parent comment.  This would probably be more
performant (avoid n^n query complexity for fetching post->comments) but
I'm going the polymorphic association route presently for simplicity.
It would be trivial to denormalize post_id into a column in comment in
the future annyway.